fix a signed/unsigned comparison warning
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@4706 d7cf8633-e32d-0410-b094-e92efae38249

+++ snd_oss.c
@@ -121,7 +121,7 @@ qboolean SNDDMA_Init(void)
 		shm->format.speed = atoi(com_argv[i+1]);
 	else
 	{
-		for (i = 0;i < (int) sizeof(tryrates) / sizeof(tryrates[0]);i++)
+		for (i = 0;i < (int) sizeof(tryrates) / (int) sizeof(tryrates[0]);i++)
 			if (!ioctl(audio_fd, SNDCTL_DSP_SPEED, &tryrates[i]))
 				break;
